First steps towards running on Linux.
Most of the .Net stuff in the Windows namespace is available in mono
as well, so it should be mostly about finding the correct decision
points, where the p/invoke versus upstart-specific stuff should be.

I've started with a vagrant instance and I've started making the project
files conditionally compile on linux, because some of the code in there
can't compile on linux at all.

More info here: http://www.mono-project.com/Porting_MSBuild_Projects_To_XBuild
, and https://github.com/ServiceStack/ServiceStack/wiki/Run-ServiceStack-as-a-daemon-on-Linux

I'm setting up a new Windows 8 environment in VirtualBox that I'm going to
use to initially build Topshelf, to see if I can use 'mono-4.0-service' to
run it.

Relates to #107.